Refactored: the assessment structure to handle question type
