+++ src/app/assessment/assessment.component.ts
@@ -14,14 +14,14 @@ import { Todo } from '../todos/todo';
         </button>
         <div *ngIf="isOpen" [ngSwitch]="assessment.type">
 
-            <div *ngSwitchCase="ASSESSMENT_TYPES.MICRO_PROJECT.id">
-                <xp-assessment-micro-project
-                    [assessment]="assessment"
-                    [todo]="todo">
-                </xp-assessment-micro-project>
-            </div>
+            <xp-assessment-micro-project
+                *ngSwitchCase="ASSESSMENT_TYPES.MICRO_PROJECT.id"
+                [assessment]="assessment"
+                [todo]="todo">
+            </xp-assessment-micro-project>
             <xp-assessment-question
-                *ngSwitchCase="ASSESSMENT_TYPES.QUESTION.id">
+                *ngSwitchCase="ASSESSMENT_TYPES.QUESTION.id"
+                [todo]="todo">
             </xp-assessment-question>
 
         </div>
