--- src/app/assessment/question/form.component.ts
@@ -1,19 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'xp-assessment-question-form',
-  template: `
-    <p>
-      assessment-question-form Works!
-    </p>
-  `,
-  styles: []
-})
-export class AssessmentQuestionFormComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
